requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.5
defusedxml==0.7.1
Django==3.2.2
django-allauth==0.58.2
djangorestframework==3.14.0
google==3.0.0
google-api-core==2.14.0
google-api-python-client==2.107.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
httplib2==0.22.0
idna==3.4
mysqlclient==2.2.0
oauthlib==3.2.2
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python3-openid==3.2.0
pytz==2023.3.post1
razorpay==1.4.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
soupsieve==2.5
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.6
